+++ WebApiTerra1000/Controllers/NomenclatureController.cs
@@ -33,7 +33,7 @@ namespace WebApiTerra1000.Controllers
         [AllowAnonymous]
         [HttpGet()]
         [Route("api/nomenclature/")]
-        public ContentResult GetNomenclature(string code, int id, FormatType format = FormatType.Xml)
+        public ContentResult GetNomenclature(string code, long id, FormatType format = FormatType.Xml)
         {
             return Controller.RunTask(new Task<ContentResult>(() =>
             {
